device/intr: Account for interrupts that could not be delivered
When an irq handler dies, we are decrementing the n_unacked count
and calling __enable_irq() the right number of times, but we need
to decrement the total interrupt count by the number that were lost
and also clear that number.

This fixes a hang when a shared irq handler quits and leaves some
unacked interrupts.

Message-ID: <20241123222020.245519-1-damien@zamaudio.com>