vm_page: Keep evicting out pages with external pager
`double_paging` used to be used to detect when we have not really flushed
a page yet (we have just pushed it to the external pager for now) and we
really want to flush something because allocations are paused. But when we
do not have a default pager, we are not double-paging but we should still
continue evicting pages.