vm_page_seg_evict: Fix creating a pager for internal objects with DMM
fd63a4bbf6f2 ("vm_page: Avoid creating a pager in DMM when not
double-paging") avoided crashes when we do not have a DMM and wish to evict
an internal opbject. But the double_paging condition was too restrictive and
prevented normal internal objects from paging out when we do have a DMM.

This re-enables creating a pager when we do have a DMM, so internal
objects can be paged out.