i386/smp.c: Change order of waiting for pending ICR
We need to wait first for pending to clear before sending
the IPI assert signal.  Likewise, if we always do it this
way, we don't need to wait for pending to clear afterward
because next time we will wait again anyway.  Deassert does
not seem necessary according to Intel SDM Vol 3, which
says an IPI is sent when ICR low is written.

NB: This is not the same code as for STARTUP or INIT IPIs.

Message-ID: <20250710101341.178343-2-damien@zamaudio.com>